Added visual feat lf-helpers;Removed unnecessary pdfinfo call
